renaming the blacklight filter (and so on) from PublicDomainFilter to CopyrightFreeFilter
