+++ app/views/application/_search_form.html.erb
@@ -47,7 +47,7 @@
     </div>
 
     <div class="form-group checkbox">
-      <label for="filter_public_domain"><%= check_box_tag 'filter_public_domain', value=1 %>Copyright Free</label>
+      <label for="filter_copyright_free"><%= check_box_tag 'filter_copyright_free', value=1 %>Copyright Free</label>
     </div>
   </div>
 <% end %>
