+++ app/views/catalog/_query_constraint_as_form.html.erb
@@ -29,6 +29,6 @@
     </div>
 </form>
 
-<% if SearchBuilder::PublicDomainFilter.filtered_public_domain?(local_assigns[:params]) %>
-  <%= render_constraint_element(nil, "Copyright Free", remove: search_state.params_for_search.except(:filter_public_domain, :page, :utf8)) %>
+<% if SearchBuilder::CopyrightFreeFilter.filtered_copyright_free?(local_assigns[:params]) %>
+  <%= render_constraint_element(nil, "Copyright Free", remove: search_state.params_for_search.except(:filter_copyright_free, :page, :utf8)) %>
 <% end %>
